This commit makes major changes in the dependencies.
It removes the MMDB2 library dependency and replaces it with the Gemmi library. This in turn requires changes in the CMake file, which now takes both the boolead USE_GEMMI variable as well as the GEMMI_PATH variable to where gemmi is installed. It also assumes that the libz library is available on the system; however, these assumptions can be removed by not using Gemmi (i.e. -DUSE_GEMMI=FALSE).
As a side-effect of this transition, it is now allowed for the mmCIF files to be gunzipped (.cif.gz) as well as uncompressed (.cif).