In Representation.cpp: The case of 1- and 3-character residue names in structure file parsing was handled ok. But DNA uses 2-character residue names, DA,DG,DC,DT. Trapped and handle d the 2-character case. Worked like a charm.
In Representation.h:
The detection of DNA and RNA was reversed. Also for the DNA(now RNA) case, was looking at the range of names ADENOSINE to THYMINE, corrected the latter to URI
DINE.

in PDBReader.cpp:
Just some debug comments. Mostly removed, left in a few that are not so frequent or verbose.